In this version i managed to handle overflow and invalid PID
also norminette friendly

+++ libft/ft_atol.c
@@ -12,7 +12,7 @@
 
 #include <limits.h>
 
-long	ft_atol(const char *nptr)
+long long	ft_atol(const char *nptr)
 {
 	long	num;
 	int		minus;
